music: merge music modules into TRX
This merges each game's music module into one in TRX.
